requirements.txt
# Google ADK and dependencies
google-adk[a2a]>=1.0.0
google-genai>=0.1.0

# A2A communication
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0

# Toolbox integration (Phase 1 - existing)
toolbox-core>=0.1.0

# Phase 2: MCP Protocol Support
fastmcp>=0.1.0
mcp>=0.1.0

# Phase 2: JWT Authentication
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# Environment management
python-dotenv>=1.0.0

# HTTP clients
requests>=2.31.0
httpx>=0.25.0

# Utilities
python-dateutil>=2.8.2
